Test out new deserialize way
