+++ ArrowExample/ArrowExample/User+JSON.swift
@@ -1,0 +1,20 @@
+//
+//  User+JSON.swift
+//  ArrowExample
+//
+//  Created by Sacha Durand Saint Omer on 13/04/16.
+//  Copyright © 2016 Sacha Durand Saint Omer. All rights reserved.
+//
+
+
+import Arrow
+
+extension User:ArrowParsable {
+    
+    public func deserialize(json: JSON) {
+        
+    }
+}
+
+
+//class = not in extension
