+++ ArrowExample/ArrowExample/User.swift
@@ -1,0 +1,14 @@
+//
+//  User.swift
+//  ArrowExample
+//
+//  Created by Sacha Durand Saint Omer on 13/04/16.
+//  Copyright © 2016 Sacha Durand Saint Omer. All rights reserved.
+//
+
+import Foundation
+
+public class User {
+
+    public required init() { }
+}
