+++ ArrowExample/ArrowTests/Stats+JSON.swift
@@ -1,12 +1,11 @@
 //
-//  Stats+Arrow.swift
-//  Swift Structs Test
+//  Stats+JSON.swift
+//  ArrowExample
 //
-//  Created by Sacha Durand Saint Omer on 6/7/15.
-//  Copyright (c) 2015 Sacha Durand Saint Omer. All rights reserved.
+//  Created by Sacha Durand Saint Omer on 13/04/16.
+//  Copyright © 2016 Sacha Durand Saint Omer. All rights reserved.
 //
 
-import Foundation
 import Arrow
 
 extension Stats:ArrowParsable {
